+++ src/core/plugins/core-commands/server/index.ts
@@ -6,26 +6,20 @@ import './commands/moderator/currency';
 import './commands/moderator/garage';
 import './commands/moderator/interior';
 import './commands/moderator/inventory';
-import './commands/moderator/item';
 import './commands/moderator/noclip';
 import './commands/moderator/player';
-import './commands/moderator/revive';
 import './commands/moderator/teleport';
 import './commands/moderator/test';
 import './commands/moderator/vehicle';
 import './commands/moderator/wanted';
 import './commands/moderator/weapon';
+import './commands/moderator/chat';
 
 import './commands/player/chat';
 import './commands/player/death';
-import './commands/player/do';
 import './commands/player/job';
-import './commands/player/low';
-import './commands/player/me';
-import './commands/player/ooc';
 import './commands/player/seatbelt';
 import './commands/player/vehicle';
-import './commands/player/whisper';
 
 const PLUGIN_NAME = 'Athena Roleplay Commands';
 PluginSystem.registerPlugin(PLUGIN_NAME, () => {
